+++ src/testFixtures/java/uk/gov/di/ipv/cri/common/library/client/CommonApiClient.java
@@ -1,8 +1,14 @@
 package uk.gov.di.ipv.cri.common.library.client;
 
+import com.nimbusds.oauth2.sdk.AuthorizationCode;
+import com.nimbusds.oauth2.sdk.AuthorizationCodeGrant;
+import com.nimbusds.oauth2.sdk.TokenRequest;
+import com.nimbusds.oauth2.sdk.auth.PrivateKeyJWT;
 import uk.gov.di.ipv.cri.common.library.util.URIBuilder;
 
 import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
@@ -10,11 +16,13 @@ import java.net.http.HttpResponse;
 public class CommonApiClient {
     private final HttpClient httpClient;
     private final ClientConfigurationService clientConfigurationService;
+    private final TestResourcesClient testResourcesClient;
 
     private static final String JSON_MIME_MEDIA_TYPE = "application/json";
 
     public CommonApiClient(ClientConfigurationService clientConfigurationService) {
         this.clientConfigurationService = clientConfigurationService;
+        this.testResourcesClient = new TestResourcesClient(clientConfigurationService);
         this.httpClient = HttpClient.newBuilder().build();
     }
 
@@ -31,6 +39,33 @@ public class CommonApiClient {
                                         .toString())
                         .addParameter(
                                 "client_id", this.clientConfigurationService.getDefaultClientId())
+                        .addParameter("response_type", "code")
+                        .addParameter("scope", "openid")
+                        .addParameter("state", "state-ipv")
+                        .build();
+
+        var request =
+                HttpRequest.newBuilder()
+                        .uri(url)
+                        .header(HttpHeaders.ACCEPT, JSON_MIME_MEDIA_TYPE)
+                        .header(HttpHeaders.SESSION_ID, sessionId)
+                        .GET()
+                        .build();
+        return sendHttpRequest(request);
+    }
+
+    public HttpResponse<String> sendNewAuthorizationRequest(String sessionId)
+            throws IOException, InterruptedException {
+        var url =
+                new URIBuilder(this.clientConfigurationService.getPrivateApiEndpoint())
+                        .setPath(this.clientConfigurationService.createUriPath("authorization"))
+                        .addParameter(
+                                "redirect_uri",
+                                new URIBuilder(testResourcesClient.getTestHarnessUrl())
+                                        .setPath("/callback")
+                                        .build()
+                                        .toString())
+                        .addParameter("client_id", "ipv-core-stub-aws-headless")
                         .addParameter("response_type", "code")
                         .addParameter("scope", "openid")
                         .addParameter("state", "state-ipv")
@@ -104,6 +139,41 @@ public class CommonApiClient {
                                 HttpHeaders.API_KEY,
                                 this.clientConfigurationService.getPublicApiKey())
                         .POST(HttpRequest.BodyPublishers.ofString(privateKeyJwt))
+                        .build();
+        return sendHttpRequest(request);
+    }
+
+    public HttpResponse<String> sendNewTokenRequest(
+            PrivateKeyJWT privateKeyJwt, String code, String issuer)
+            throws IOException, InterruptedException, URISyntaxException {
+        var authorisationGrant =
+                new AuthorizationCodeGrant(
+                        new AuthorizationCode(code), new URI(issuer + "/callback"));
+        var tokenRequest =
+                new TokenRequest(
+                        new URIBuilder(this.clientConfigurationService.getPublicApiEndpoint())
+                                .setPath(this.clientConfigurationService.createUriPath("token"))
+                                .build(),
+                        privateKeyJwt,
+                        authorisationGrant);
+
+        var tokenRequestBody = tokenRequest.toHTTPRequest();
+
+        var request =
+                HttpRequest.newBuilder()
+                        .uri(
+                                new URIBuilder(
+                                                this.clientConfigurationService
+                                                        .getPublicApiEndpoint())
+                                        .setPath(
+                                                this.clientConfigurationService.createUriPath(
+                                                        "token"))
+                                        .build())
+                        .header(HttpHeaders.CONTENT_TYPE, "application/x-www-form-urlencoded")
+                        .header(
+                                HttpHeaders.API_KEY,
+                                this.clientConfigurationService.getPublicApiKey())
+                        .POST(HttpRequest.BodyPublishers.ofString(tokenRequestBody.getBody()))
                         .build();
         return sendHttpRequest(request);
     }